+++ src/pages/home/projects/Emociones.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { ExpandableSection } from 'components';
 
-const Emociones = ({ openTab, ...props }) => {
+const Emociones = ({ id, openTab, ...props }) => {
     return (
-        <ExpandableSection {...props} id="emociones" title="Emociones" openTab={openTab}>
+        <ExpandableSection {...props} id={id} title="Emociones" openTab={openTab}>
             <></>
         </ExpandableSection>
     );