+++ src/pages/home/projects/Malamiga.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { ExpandableSection } from 'components';
 
-const Malamiga = ({ openTab, ...props }) => {
+const Malamiga = ({ id, openTab, ...props }) => {
     return (
-        <ExpandableSection {...props} id="malamiga" title="Malamiga" openTab={openTab}>
+        <ExpandableSection {...props} id={id} title="Malamiga" openTab={openTab}>
             <></>
         </ExpandableSection>
     );